requirements.txt
selenium
requests
pytest
allure-pytest
python-dotenv